+++ Nickvision.Aura/Configuration/ConfigurationLoader.cs
@@ -3,12 +3,12 @@ using System.Text.Encodings.Web;
 using System.Text.Json;
 using System.Text.Unicode;
 
-namespace Nickvision.Aura;
+namespace Nickvision.Aura.Configuration;
 
 /// <summary>
 /// Loader for JSON configuration files
 /// </summary>
-public static class ConfigurationLoader
+internal static class ConfigurationLoader
 {
     /// <summary>
     /// Load object from JSON file
@@ -16,7 +16,7 @@ public static class ConfigurationLoader
     /// <typeparam name="T">Type of the object to deserialize</typeparam>
     /// <param name="key">File name</param>
     /// <returns>Loaded or new object</returns>
-    internal static T Load<T>(string key) where T : ConfigurationBase
+    public static T Load<T>(string key) where T : ConfigurationBase
     {
         var path = $"{UserDirectories.ApplicationConfig}{Path.DirectorySeparatorChar}{key}.json";
         try
@@ -39,7 +39,7 @@ public static class ConfigurationLoader
     /// </summary>
     /// <param name="obj">IConfiguration object to save</param>
     /// <param name="key">File name</param>
-    internal static void Save(ConfigurationBase obj, string key)
+    public static void Save(ConfigurationBase obj, string key)
     {
         File.WriteAllText($"{UserDirectories.ApplicationConfig}{Path.DirectorySeparatorChar}{key}.json", JsonSerializer.Serialize((object)obj, new JsonSerializerOptions { WriteIndented = true, Encoder = JavaScriptEncoder.Create(UnicodeRanges.All) }));
         obj.RaiseSavedEvent();
